Started reorganizing file structure
