+++ server/.eslintrc.js
@@ -1,0 +1,19 @@
+module.exports = {
+  env: {
+    es2020: true,
+  },
+  extends: [
+    'plugin:vue/essential',
+    'airbnb-base',
+  ],
+  parserOptions: {
+    ecmaVersion: 11,
+    sourceType: 'module',
+  },
+  plugins: [
+    'vue',
+  ],
+  rules: {
+    'linebreak-style': 0,
+  },
+};
