--- server/config/database.js
@@ -1,7 +1,0 @@
-const mongoose = require('mongoose');
-require('dotenv').config();
-
-mongoose.connect(process.env.DATABASE_URL, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-});
