+++ server/config/default.json
@@ -1,0 +1,5 @@
+{
+  "user": {
+    "saltWorkValue": 10
+  }
+}
