+++ server/users/errors.js
@@ -1,0 +1,13 @@
+module.exports.notFound = (req, res, next) => {
+  res.status(404);
+  const error = new Error(`Not Found - ${req.originalUrl}`);
+  next(error);
+};
+
+module.exports.handleError = (err, req, res, next) => {
+  res.status(res.statusCode || 500);
+  res.json({
+    message: err.message,
+    stack: process.env.NODE_ENV === 'production' ? err.stack : 'hehe',
+  });
+};
